requirements.txt
meteoredpy==0.0.4
discord==1.7.3
discord.py==1.7.3
python_dotenv==1.0.0
mysql==0.0.3